Move initialization from Master.onSync
